Move QueueSet out of BoundsAnalysis.h
The class QueueSet was added to BoundsAnalysis.h to enable processing unique
elements in the FIFO order. However, this class is generic enough and can be
used in other places. So we are moving it out of BoundsAnalysis.h
